+++ clang-tools-extra/clang-tidy/utils/PureUtils.cpp
@@ -19,7 +19,7 @@ namespace pure {
 using utils::lexer::getTokenKind;
 using utils::lexer::forwardSkipWhitespaceAndComments;
   
-static SourceLocation checkLocationForFix(SourceLocation TargetLoc, const ASTContext &Context, const SourceManager &SM) {
+static SourceLocation checkLocationForFix(SourceLocation TargetLoc, const ASTContext &Context, const SourceManager &SM, const char *TraceMethod) {
   if (!TargetLoc.isValid())
     return TargetLoc;
 
@@ -30,13 +30,13 @@ static SourceLocation checkLocationForFix(SourceLocation TargetLoc, const ASTCon
   const char *Begin = SM.getCharacterData(TargetLoc);
   const char *End = SM.getCharacterData(EndLoc);
   llvm::StringRef TestToken(Begin, End - Begin);
-  if (!TestToken.compare("trace")) // fix apparently already applied
+  if (!TestToken.compare(TraceMethod)) // fix apparently already applied
     return SourceLocation();
 
   return TargetLoc;
 }
 
-SourceLocation condSkipDeclaration(SourceLocation Loc, const ASTContext &Context) {
+SourceLocation condSkipDeclaration(SourceLocation Loc, const ASTContext &Context, const char *TraceMethod) {
   if (!Loc.isValid())
     return Loc;
 
@@ -46,10 +46,10 @@ SourceLocation condSkipDeclaration(SourceLocation Loc, const ASTContext &Context
 
   const SourceManager &SM = Context.getSourceManager();
   SourceLocation EndLoc = forwardSkipWhitespaceAndComments(NextLoc, SM, &Context);
-  return checkLocationForFix(EndLoc, Context, SM);
+  return checkLocationForFix(EndLoc, Context, SM, TraceMethod);
 }
 
-SourceLocation condSkipStatement(SourceLocation Loc, const ASTContext &Context) {
+SourceLocation condSkipStatement(SourceLocation Loc, const ASTContext &Context, const char *TraceMethod) {
   SourceLocation InvalidLoc;
   if (!Loc.isValid())
     return InvalidLoc;
@@ -65,14 +65,14 @@ SourceLocation condSkipStatement(SourceLocation Loc, const ASTContext &Context) 
 
   tok::TokenKind TokKind = getTokenKind(SemiLoc, SM, &Context);
   if (TokKind != tok::semi) // happens for if statements
-    return checkLocationForFix(SemiLoc, Context, SM);
+    return checkLocationForFix(SemiLoc, Context, SM, TraceMethod);
 
   SourceLocation AfterLoc = Lexer::getLocForEndOfToken(SemiLoc, 0, SM, Context.getLangOpts());
   if (!AfterLoc.isValid())
     return InvalidLoc;
 
   SourceLocation EndLoc = AfterLoc.getLocWithOffset(1);
-  return checkLocationForFix(EndLoc, Context, SM);
+  return checkLocationForFix(EndLoc, Context, SM, TraceMethod);
 }
 
 std::string makeTraceFix(llvm::StringRef BufferName, int HexFlag) {
@@ -91,7 +91,18 @@ std::string makeTraceFix(llvm::StringRef BufferName, int HexFlag) {
   Fix += ");\n";
   return Fix;
 }
-  
+
+std::string makeTraceFloatFix(llvm::StringRef BufferName) {
+  std::string Fix("trace_num((uint32_t)\"");
+  Fix += BufferName;
+  Fix += "\", sizeof(\"";
+  Fix += BufferName;
+  Fix += "\"), (int64_t)";
+  Fix += BufferName;
+  Fix += ");\n";
+  return Fix;
+}
+
 } // namespace pure
 } // namespace utils
 } // namespace tidy